requirements/base.txt
# Framework
Django==2.2.1

# Better configuration
django-configurations==2.1

# Postgres
psycopg2==2.7.3.1
dj-database-url==0.5.0

# Tests
pytest==4.5.0
pytest-cov==2.7.1
pytest-django==3.4.8
